Rename testing files. More easy to search
